Temporary responses.coffee anti-spam update
With the intention of slowing down some of the most common responses, I think this a good temporary fix. If people really want, they can comeback and re-add them. But the key purpose is to de-clutter in an attempt to remove spam. I kept some of the rarer ones I've never seen called, and indiscriminately removed the others.